+++ src/test/java/config/CredentialConfig.java
@@ -1,0 +1,12 @@
+package config;
+
+import org.aeonbits.owner.Config;
+
+@Config.Sources("classpath:config/credentials.properties")
+
+public interface  CredentialConfig extends Config {
+
+    String login();
+    String password();
+    String remoteUrl();
+}
